fix: 버그 및 mysql 트리거, 함수 수정 (#62)
* fix: 시리즈테이블 트리거 수정

게시글이 추가되었을 때 시리즈의 썸네일이 null일 경우 해당 포스터에 썸네일로 업데이트 하도록 수정

* fix: 시리즈 게시글의 정렬 순서 함수 수정

시리즈 게시글의 정렬 순서를 가지고 오는 함수 수정

* fix: 게시글 수정 시 시리즈 추가 오류 수정

게시글 수정 시에 시리즈가 포함된 경우 중복적으로 시리즈에 추가되는 오류 수정

+++ src/post/post.service.ts
@@ -107,18 +107,20 @@ export class PostService {
       await this.createTag(data.tags, post_id);
     }
 
-    if (data.series_id == null) {
-      const post_series = await this.postSeriesRepository.getPostSeriesId(post_id);
+    const series_id = await this.postSeriesRepository.getPostSeries(post_id);
 
+    if (!series_id && data.series_id) {
+      await this.postSeriesRepository.createPostSeries(post_id, data.series_id);
+    }
+
+    if (series_id && data.series_id == null) {
       await this.postSeriesRepository.deletePostSeries(post_id, null);
 
-      if (post_series.length != 0) {
-        for (let i = 0; i < post_series.length; i++) {
-          await this.postSeriesRepository.updateSort(i + 1, post_series[i].id);
-        }
+      const post_series_id = await this.postSeriesRepository.getPostSeriesId(series_id.series_id);
+
+      for (let i = 0; i < post_series_id.length; i++) {
+        await this.postSeriesRepository.updateSort(i + 1, post_series_id[i].id);
       }
-    } else {
-      await this.postSeriesRepository.createPostSeries(post_id, data.series_id);
     }
 
     return post_id;
